2025_LLM_Agents_course_and_develpoe/Audio_Agent_meeting_recorder/MeetingRecorderAI/AI_meeting_by_Gradio/requirements.txt
gradio>=5.0.0
pyaudio>=0.2.13
openai>=1.0.0
numpy>=1.20.0
python-dotenv>=1.0.0
